initial implementation of cross promotion
